Rename records.hints to records.delimited
