+++ records_mover/records/job/config.py
@@ -5,7 +5,7 @@ from records_mover import Session
 from ..records_format import DelimitedRecordsFormat
 from ..existing_table_handling import ExistingTableHandling
 from .hints import SUPPORTED_HINT_LOOKUP
-from ..hints import RecordsHints
+from ..delimited import RecordsHints
 from ...mover_types import JobConfig
