+++ records_mover/records/pandas/to_csv_options.py
@@ -1,6 +1,6 @@
 import csv
 from ...utils import quiet_remove
-from ..hints import cant_handle_hint
+from ..delimited import cant_handle_hint
 from ..processing_instructions import ProcessingInstructions
 from ..records_format import DelimitedRecordsFormat
 import logging
