+++ records_mover/records/records_format.py
@@ -3,7 +3,7 @@ from .processing_instructions import ProcessingInstructions
 from . import RecordsHints
 from .base_records_format import BaseRecordsFormat
 from typing import Mapping, Optional, TYPE_CHECKING
-from .hints import MutableRecordsHints, ValidatedRecordsHints
+from .delimited import MutableRecordsHints, ValidatedRecordsHints
 if TYPE_CHECKING:
     from . import RecordsFormatType  # noqa
