+++ tests/unit/records/test_csv_streamer.py
@@ -1,11 +1,11 @@
 import unittest
 from mock import patch, Mock
-from records_mover.records.hints import stream_csv
+from records_mover.records.delimited import stream_csv
 
 
 @patch('pandas.read_csv')
 @patch('pandas.io.parsers.TextFileReader')
-@patch('records_mover.records.hints.csv_streamer.io')
+@patch('records_mover.records.delimited.csv_streamer.io')
 class TestCsvStreamer(unittest.TestCase):
     def test_stream_csv_no_compression(self,
                                        mock_io,
