+++ tests/unit/records/test_hints.py
@@ -1,4 +1,4 @@
-from records_mover.records.hints.sniff import (
+from records_mover.records.delimited.sniff import (
     sniff_hints, sniff_hints_from_fileobjs, sniff_encoding_hint, BootstrappingRecordsHints
 )
 from mock import MagicMock, patch
@@ -35,8 +35,8 @@ class TestHints(unittest.TestCase):
                                 f"Expected at least these hints while reading {basename}: "
                                 f"{required_hints}, found these hints: {hints}")
 
-    @patch('records_mover.records.hints.sniff.stream_csv')
-    @patch('records_mover.records.hints.sniff.io')
+    @patch('records_mover.records.delimited.sniff.stream_csv')
+    @patch('records_mover.records.delimited.sniff.io')
     def test_sniff_hints_from_fileobjs(self,
                                        mock_io,
                                        mock_stream_csv) -> None:
